Added gtest library, refactored makefile
Now we can run `make test` to create test excecutables.
Note that we create all the .o files in the currect directory,
the alternative is to launch make from the build directory,
or prefix all the left hand sides of .o rules with $(BUILD).
See http://mad-scientist.net/make/vpath.html
and
http://mad-scientist.net/make/multi-arch.html