Tested CNOT_synth
I finally figured out which was was which for the target and control
lines.
I also had to debug a implicit promotion from 4 to xor_func{4},
which was causing vector<xor_func> v{4, xor_func{{false, {1,1,1,1}}}};
to create a vector of size 2, not of size 4.
Still todo is understanding how m is picked in CNOT_synth.